+++ src/CommandLine/Rename/CliCompilationErrorResolution.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Josef Pihrt and Contributors. Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Roslynator.Rename;
+namespace Roslynator.CommandLine.Rename;
 
-internal enum RenameErrorResolution
+internal enum CliCompilationErrorResolution
 {
     None = 0,
     Abort = 1,